Makefile and fixes.
- Added makefile.
- Fixed all warnings generated by gcc.

+++ src/token.c
@@ -28,7 +28,7 @@ void token_list_json_dump(FILE* stream, list_t* list)
     fprintf(stream, "{\"kind\":\"%s\"", token_kind_to_string(tok->kind));
 
     if (token_is_literal(tok) || tok->kind == TOK_ID)
-      fprintf(stream, ",\"value\":\"%.*s\"", tok->loc->len, tok->loc->cur);
+      fprintf(stream, ",\"value\":\"%.*s\"", (int)tok->loc->len, tok->loc->cur);
 
     fputc('}', stream);
